Add a --sudo flag to api-cli (#102)
* Add a --sudo flag to api-cli

Suggested in the comments to #100, this enables CLI users to send
transactions which require superuser authentication. Permission-
raising code adapted from https://polkadot.js.org/api/examples/promise/10_upgrade_chain/.

There's a bit of duck-typing added in `makeTx`, largely because it was
very unclear what interfaces are actually appropriate for `signable`
and `auth`. `ApiCallResult` doesn't work for both code branches.
Additionally, casting `fn(...params) as unknown as Proposal` feels
like it may not be the best possible approach, though it's not clear
what a better one would be.

* cast return type to appease the typechecker

* Require that the user specify the seed while sudoing

This removes the dependency on the test keyring, and should ensure
that sudo functionality works for all chains which permit sudo
in the first place.

* Update sudo command description

Co-Authored-By: Jaco Greeff <jacogr@gmail.com>

* Propagate the original API object instead of recreating on demand

This avoids a second invocation of a fairly heavy call to construct
the API.

We also use constant auth information, and put the responsibility
on the user to supply the correct sudo key.

Co-authored-by: Jaco Greeff <jacogr@gmail.com>